move new test to test_io - rename to from_callable
