+++ dask/dataframe/__init__.py
@@ -18,7 +18,7 @@ try:
         read_hdf,
         read_sql_table,
         from_delayed,
-        from_function,
+        from_callable,
         read_csv,
         to_csv,
         read_table,
