+++ dask/dataframe/io/tests/test_io.py
@@ -7,6 +7,7 @@ from multiprocessing.pool import ThreadPool
 
 import dask.array as da
 import dask.dataframe as dd
+from dask.blockwise import Blockwise, optimize_blockwise
 from dask.dataframe._compat import tm
 from dask.dataframe.io.io import _meta_from_array
 from dask.delayed import Delayed, delayed
@@ -697,3 +698,53 @@ def test_from_dask_array_index_dtype():
 
     assert ddf.index.dtype == ddf2.index.dtype
     assert ddf.index.name == ddf2.index.name
+
+
+def test_from_callable():
+
+    # Function used to generate data for each partition
+    def _create(part, size):
+        return pd.DataFrame({"a": np.arange(part * size, (part + 1) * size)})
+
+    npartitions = 4  # Output partition count
+    size = 10  # Argument to `_create`
+
+    # Use `from_callable` to generate a DataFrame
+    # comprised of a single `Blockwise` layer
+    ddf = dd.from_callable(_create, npartitions, size)
+    expect = _create(0, size * npartitions)
+
+    # `ddf` should now have ONE Blockwise layer
+    layers = ddf.__dask_graph__().layers
+    assert len(layers) == 1
+    assert isinstance(list(layers.values())[0], Blockwise)
+
+    # Check single-layer result
+    got = ddf.copy().compute().reset_index(drop=True)
+    assert_eq(got, expect.reset_index(drop=True))
+
+    # Increment by 1
+    ddf += 1
+    expect += 1
+
+    # Increment by 10
+    ddf += 10
+    expect += 10
+
+    # `ddf` should now have THREE Blockwise layers
+    layers = ddf.__dask_graph__().layers
+    assert len(layers) == 3
+    assert all([isinstance(layer, Blockwise) for (_, layer) in layers.items()])
+
+    # Check that `optimize_blockwise` fuses all three
+    # `Blockwise` layers together
+    keys = [(ddf._name, i) for i in range(npartitions)]
+    graph = optimize_blockwise(ddf.__dask_graph__(), keys)
+    layers = graph.layers
+    name = list(layers.keys())[0]
+    assert len(layers) == 1
+    assert isinstance(layers[name], Blockwise)
+
+    # Check final result
+    got = ddf.compute().reset_index(drop=True)
+    assert_eq(got, expect.reset_index(drop=True))
